+++ go.mod
@@ -12,7 +12,7 @@ require (
 	github.com/cheynewallace/tabby v1.1.1
 	github.com/cloudnative-pg/barman-cloud v0.0.0-20241105055149-ae6c2408bd14
 	github.com/cloudnative-pg/cnpg-i v0.0.0-20241105133936-c704f46c20e0
-	github.com/cloudnative-pg/machinery v0.0.0-20241121132610-4ee66a4c3b4f
+	github.com/cloudnative-pg/machinery v0.0.0-20241122084004-33b997fc6c61
 	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc
 	github.com/evanphx/json-patch/v5 v5.9.0
 	github.com/go-logr/logr v1.4.2
