+++ pkg/management/postgres/constants/constants.go
@@ -51,4 +51,9 @@ const (
 
 	// Startup is the name of a file that is created once during the first reconcile of an instance
 	Startup = "cnpg_initialized"
+
+	// CheckEmptyWalArchiveFile is the name of the file in the PGDATA that,
+	// if present, requires the WAL archiver to check that the backup object
+	// store is empty.
+	CheckEmptyWalArchiveFile = ".check-empty-wal-archive"
 )
